feat: allow default arguments to be specified
Because of the declarative nature of clippy, we can't know what the type
of the default argument should be. I tried various things using
`?*const anyopaque` as a way of representing any value, but in the end
it seemed easier just to have everything be specified as strings, and
then we compile time parse them.