feat!: allow for fallback commands in Command
The use case is for commands where there are a set of defined outcomes,
but then also a number of wildcard alternatives. E.g., "a" and "b" are
valid commands with subcommands, but any other string, e.g. "c" should
be treated seperately as it's own command.

This breaks the current API in that commands now take ArgumentDescriptor
lists instead of Arguments. Ultimately, however, this lends more type
stability to the library.

There is an outstanding issue in that the help string is currently
erroneous, in that there is a bug where the fallback command and it's
internal argument are both printed.