feat: command alises
Can now pass an alias string hash map to the init of Commands, which
will perform string substitution when evaluating what the command is. It
cannot be used to perform full aliasing for flags etc., but can be used
to make e.g. `l`, `ls`, and `lis` all resolve to the `list` command.